Working Dice roller test
The dice roller test is now working.
Rolls 20 dice and packs the rolls in the results array
Includes utility functions to unpack the result

Issues:
Currnently, the rolls are timed using deterministically timed key presses. Rolls seem to be random when run on demoboard. Perhaps randomization is needed on the satellite platform.